20240125 删除多余代码，能跑起来

+++ TMessagesProj/src/main/java/org/telegram/ui/Components/ChatActivityInterface.java
@@ -37,7 +37,7 @@ public interface ChatActivityInterface {
         return 0;
     }
 
-    default long getTopicId() {
+    default int getTopicId() {
         return 0;
     }
